packages/init-scripts/pyproject.toml
[project]
name = "init-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Kuroki-g", email = "64629990+Kuroki-g@users.noreply.github.com" }
]
requires-python = "==3.13.*"
dependencies = [
    "common",
    "mistune>=3.2.0",
    "pdfminer-six>=20260107",
]

[project.scripts]
init = "init_scripts:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
common = { workspace = true }
